HS-380 Refact tests
