+++ ui/tests/store/Queries/appliancesQueries.test.ts
@@ -1,7 +1,7 @@
 import { createTestingPinia } from '@pinia/testing'
 import { useApplianceQueries } from '@/store/Queries/applianceQueries'
 import { minionsFixture, expectedAppliancesMinions } from '../../fixture/minions'
-import { devicesFixture, expectedAppliancesDevices } from '../../fixture/devices'
+import { devicesFixture } from '../../fixture/devices'
 import { minionLatencyFixture, minionUptimeFixture, deviceLatencyFixture, deviceUptimeFixture } from '../../fixture/metrics'
 import { locationsFixture, expectedLocations } from '../../fixture/locations'
 
@@ -30,6 +30,18 @@ describe('Appliances queries', () =>{
         } 
       }))
     }))
+
+    const expectedAppliancesDevices = [
+      {
+        id: 1,
+        label: 'France',
+        createTime: '2022-09-07T17:52:51Z',
+        managementIp: '127.0.0.1',
+        icmp_latency: 0.17,
+        snmp_uptime: undefined,
+        status: 'DOWN'
+      }
+    ]    
 
     const appliancesQueries = useApplianceQueries()
     expect(appliancesQueries.tableDevices).toStrictEqual(expectedAppliancesDevices)
